+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PlaylistMaker"
+        android:name="com.example.playlistmaker.App"
         tools:targetApi="31">
         <activity
             android:name=".player.PlayerActivity"
